+++ cmd/blast-donation/main.go
@@ -47,7 +47,7 @@ type FieldMap map[string]*Stats
 //All reads all of the records and sends them to a Fields channel.
 //parses the buffer for records then outputs them to cout.
 func All(t *godig.Table, crit string, cout chan Fields) {
-	offset := 0
+	offset := int32(0)
 	count := 500
 	for count > 0 {
 		log.Printf("All: %v offset %6d\n", t.Name, offset)
@@ -70,7 +70,7 @@ func All(t *godig.Table, crit string, cout chan Fields) {
 		for _, r := range a {
 			cout <- r
 		}
-		offset = offset + count
+		offset = offset + int32(count)
 	}
 }
